requirements.txt
pyNaCl~=1.4.0
discord~=1.0.1
discord-ext-typed-commands
threaded
asyncio
python-dotenv